[driver/opc] refactor files
