+++ driver/opc/BUILD.bazel
@@ -5,11 +5,11 @@ cc_library(
         "reader.cpp",
         "scanner.cpp",
         "util.cpp",
-        "task.cpp",
-        "writerSink.cpp",
+        "writer_task.cpp",
+        "writer_sink.cpp",
     ],
     hdrs = [
-        "opc.h",
+        "opc_pub.h",
         "reader.h",
         "scanner.h",
         "util.h",
