+++ driver/opc/writer.h
@@ -11,7 +11,7 @@
 
 #include <map>
 
-#include "opc.h"
+#include "opc_pub.h"
 #include "util.h"
 #include "driver/config/config.h"
 #include "driver/task/task.h"
