+++ driver/opc/writer_sink.cpp
